requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
dj-database-url
psycopg2-binary
gunicorn
python-dotenv
whitenoise
channels==4.0.0
channels-redis==4.1.0
django-redis==5.4.0
